requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0
Pillow>=10.0.0
